Fixed biased averaging in parse_svm_output
While k-mers are generated encompassing corresponding characters of
original sequence in a symmetric window, the denoising step in
prediction used a sliding window that assumed that the character that
the window represents is the first one in the window. This shifted the
beginning of a region: it started k/2 characters earlier than it should.
Now the averaging is fixed so that the sliding window is symmetric
around the pivot character just as in data generation step.